Move url helpers to own file
