+++ bundles/mapping/mapmodule/plugin/getinfo/ValueFormatters.js
@@ -1,4 +1,4 @@
-
+import { shortenUrl } from './ValueFormattersUrlHelpers';
 // ----------------------------------------------------------------------
 // Built-in formatters
 // TODO: add params handling for formatters
@@ -14,55 +14,6 @@ const linkFormatter = (value, params = {}) => {
         }
     }
     return `<a href="${value}" rel="noreferrer noopener" target="_blank" title="${value}">${label}</a>`;
-};
-
-const WELL_KNOWN_PROTO = ['http://', 'https://'];
-const removeWellknownProtocolFromUrl = (url = '') => {
-    // this function is just to handle url in case-insensitive way
-    if (!url) {
-        return url;
-    }
-    const input = url.toLowerCase();
-    const foundProto = WELL_KNOWN_PROTO.findIndex(proto => input.startsWith(proto));
-    if (foundProto === -1) {
-        return url;
-    }
-    return url.substring(WELL_KNOWN_PROTO[foundProto].length);
-};
-
-const shortenUrl = (url = '', maxLength = 50) => {
-    if (url.length <= maxLength) {
-        return url;
-    }
-    const urlValue = removeWellknownProtocolFromUrl(url);
-    if (urlValue.length <= maxLength) {
-        return urlValue;
-    }
-    const partLength = maxLength / 2;
-    const start = shortenString(urlValue, partLength);
-    const end = reverseString(shortenString(reverseString(urlValue), partLength));
-    return start + '...' + end;
-};
-
-const reverseString = (input = '') => input.split('').reverse().join('');
-
-const SHORTEN_STOP_CHARS = [' ', '/', '&', '?'];
-const shortenString = (input = '', maxLength = 25) => {
-    if (input.length < maxLength) {
-        return input;
-    }
-    const threshold = maxLength * 0.8;
-    let candidate = input.substring(0, threshold);
-    let restOfInput = input.substring(threshold).split('');
-    while (restOfInput.length && candidate.length < maxLength) {
-        const character = restOfInput.shift();
-        if (SHORTEN_STOP_CHARS.includes(character)) {
-            // encountered char that is natural shorten point
-            break;
-        }
-        candidate += character;
-    }
-    return candidate;
 };
 
 const imgFormatter = (value, params = {}) => {
